gcp/requirements.txt
tensorflow==2.16.2
google-cloud-storage==504.0.1
Pillow==6.0.0